chore: move the sync status from vredential list to credential detail
